Project/requirements_ensemble.txt
# Ensemble Audio Emotion Analysis Requirements
# Extends the base requirements.txt with additional dependencies

# Base ML and Audio Processing (from existing requirements.txt)
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
librosa>=0.10.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
soundfile>=0.12.0
scipy>=1.10.0

# Additional requirements for ensemble model
# Speech-to-Text (Whisper)
openai-whisper>=20231117

# Text emotion analysis (BERT)
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.12.0
tokenizers>=0.13.0

# Additional audio processing
ffmpeg-python>=0.2.0

# Optional but recommended for better performance
# CUDA support (if available)
# torch[cuda] 

# For faster tokenization
sentencepiece>=0.1.99

# For better audio format support
pydub>=0.25.1
